requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.3
requests==2.31.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
